backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
pandas>=1.3.0
ultralytics>=8.0.0  # for YOLOv5
pytest>=7.0.0  # for testing
